backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlmodel==0.0.19
python-multipart==0.0.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
pydantic-settings==2.2.1
python-dotenv==1.0.1
torch>=2.1.0
torchvision>=0.16.0
Pillow>=9.5.0
